Update test for capybara
