+++ spec/examples.txt
@@ -1,13 +1,50 @@
-example_id                            | status | run_time        |
-------------------------------------- | ------ | --------------- |
-./spec/features/user_spec.rb[1:1:1]   | passed | 0.00293 seconds |
-./spec/features/user_spec.rb[1:1:2]   | passed | 0.02537 seconds |
-./spec/features/user_spec.rb[1:1:3]   | passed | 0.1687 seconds  |
-./spec/features/user_spec.rb[1:1:4]   | passed | 0.00869 seconds |
-./spec/features/user_spec.rb[1:1:5:1] | passed | 0.0114 seconds  |
-./spec/features/user_spec.rb[1:1:5:2] | passed | 0.12157 seconds |
-./spec/features/user_spec.rb[1:1:5:3] | passed | 0.03214 seconds |
-./spec/features/user_spec.rb[1:1:5:4] | passed | 0.00795 seconds |
-./spec/features/user_spec.rb[1:1:5:5] | passed | 0.01035 seconds |
-./spec/features/user_spec.rb[1:1:5:6] | passed | 0.01222 seconds |
-./spec/features/user_spec.rb[1:1:5:7] | passed | 0.03071 seconds |
+example_id                                | status | run_time        |
+----------------------------------------- | ------ | --------------- |
+./spec/controllers/posts_spec.rb[1:1:1:1] | failed | 0.29019 seconds |
+./spec/controllers/posts_spec.rb[1:1:2:1] | failed | 0.00911 seconds |
+./spec/controllers/posts_spec.rb[1:2:1:1] | failed | 0.00668 seconds |
+./spec/controllers/posts_spec.rb[1:2:2:1] | failed | 0.00596 seconds |
+./spec/controllers/posts_spec.rb[1:3:1:1] | failed | 0.00597 seconds |
+./spec/controllers/posts_spec.rb[1:3:2:1] | failed | 0.01133 seconds |
+./spec/controllers/users_spec.rb[1:1:1]   | failed | 0.02471 seconds |
+./spec/controllers/users_spec.rb[1:2:1]   | failed | 0.01011 seconds |
+./spec/controllers/users_spec.rb[1:3:1]   | failed | 0.00943 seconds |
+./spec/features/post_index_spec.rb[1:1:1] | passed | 0.23803 seconds |
+./spec/features/post_index_spec.rb[1:1:2] | passed | 0.02317 seconds |
+./spec/features/post_index_spec.rb[1:1:3] | passed | 0.06985 seconds |
+./spec/features/post_index_spec.rb[1:1:4] | passed | 0.02148 seconds |
+./spec/features/post_index_spec.rb[1:1:5] | passed | 0.01898 seconds |
+./spec/features/post_index_spec.rb[1:1:6] | passed | 0.02904 seconds |
+./spec/features/post_index_spec.rb[1:1:7] | passed | 0.06193 seconds |
+./spec/features/post_index_spec.rb[1:1:8] | passed | 0.01928 seconds |
+./spec/features/post_show_spec.rb[1:1:1]  | passed | 0.02981 seconds |
+./spec/features/post_show_spec.rb[1:1:2]  | failed | 0.04511 seconds |
+./spec/features/post_show_spec.rb[1:1:3]  | passed | 0.02763 seconds |
+./spec/features/post_show_spec.rb[1:1:4]  | passed | 0.0346 seconds  |
+./spec/features/post_show_spec.rb[1:1:5]  | passed | 0.04667 seconds |
+./spec/features/post_show_spec.rb[1:1:6]  | passed | 0.0283 seconds  |
+./spec/features/post_show_spec.rb[1:1:7]  | passed | 0.0388 seconds  |
+./spec/features/user_index_spec.rb[1:1:1] | passed | 0.0208 seconds  |
+./spec/features/user_index_spec.rb[1:1:2] | passed | 0.01795 seconds |
+./spec/features/user_index_spec.rb[1:1:3] | passed | 0.02419 seconds |
+./spec/features/user_index_spec.rb[1:1:4] | failed | 0.0444 seconds  |
+./spec/features/user_show_spec.rb[1:1:1]  | passed | 0.02621 seconds |
+./spec/features/user_show_spec.rb[1:1:2]  | passed | 0.02442 seconds |
+./spec/features/user_show_spec.rb[1:1:3]  | passed | 0.02761 seconds |
+./spec/features/user_show_spec.rb[1:1:4]  | passed | 0.03494 seconds |
+./spec/features/user_show_spec.rb[1:1:5]  | passed | 0.02463 seconds |
+./spec/features/user_show_spec.rb[1:1:6]  | passed | 0.02365 seconds |
+./spec/features/user_show_spec.rb[1:1:7]  | passed | 0.03655 seconds |
+./spec/models/comment_spec.rb[1:1]        | passed | 0.00187 seconds |
+./spec/models/like_spec.rb[1:1]           | passed | 0.00113 seconds |
+./spec/models/like_spec.rb[1:2]           | passed | 0.00141 seconds |
+./spec/models/post_spec.rb[1:1]           | passed | 0.0036 seconds  |
+./spec/models/post_spec.rb[1:2]           | passed | 0.00213 seconds |
+./spec/models/post_spec.rb[1:3]           | passed | 0.00218 seconds |
+./spec/models/post_spec.rb[1:4]           | passed | 0.00215 seconds |
+./spec/models/post_spec.rb[1:5]           | passed | 0.00222 seconds |
+./spec/models/user_spec.rb[1:1]           | failed | 0.26713 seconds |
+./spec/models/user_spec.rb[1:2]           | passed | 0.0017 seconds  |
+./spec/models/user_spec.rb[1:3]           | passed | 0.002 seconds   |
+./spec/models/user_spec.rb[1:4]           | passed | 0.0016 seconds  |
+./spec/models/user_spec.rb[1:5]           | failed | 0.00171 seconds |
